implementation of dynamic to review list
